fix roc_curve logic in harbor
